Demo page

+++ backend/controllers/sceneRouter.js
@@ -1,21 +1,35 @@
 const sceneRouter = require('express').Router(); // eslint-disable-line new-cap
 const jwt = require('jsonwebtoken');
 const Scene = require('../models/sceneModel');
-const {SECRET_ADMIN, SECRET_ENTER} = require('../utils/config');
+const {
+  SECRET_ADMIN,
+  SECRET_ADMIN_DEMO,
+  SECRET_ENTER,
+  SECRET_ENTER_DEMO,
+} = require('../utils/config');
 
 // creating new scene (requires ADMIN token)
 sceneRouter.post('/', async (request, response, next) => {
-  const {sceneName} = request.body;
+  const isDemo = request.isDemo;
+  let {sceneName} = request.body;
 
   // token bits
-  const adminTokenFound = jwt.verify(request.token, SECRET_ADMIN);
+  const adminTokenFound = request.isDemo ?
+    jwt.verify(request.token, SECRET_ADMIN_DEMO) :
+    jwt.verify(request.token, SECRET_ADMIN);
 
   if (!adminTokenFound) {
     return response.status(401).json({error: 'valid admin token required'});
   }
 
+  // avoid duplicate entries in the DB
+  if (isDemo) {
+    sceneName = `${sceneName}-demo`;
+  }
+
   const scene = new Scene({
     sceneName,
+    isDemo,
   });
 
   const savedScene = await scene.save();
@@ -27,12 +41,15 @@ sceneRouter.post('/', async (request, response, next) => {
 
 // getting all scenes (requires ENTRY token)
 sceneRouter.get('/', async (request, response, next) => {
-  const entryTokenFound = jwt.verify(request.token, SECRET_ENTER).id;
+  const isDemo = request.isDemo;
+  // 'entry-demo' creation uses different secret
+  const entrySecret= isDemo ? SECRET_ENTER_DEMO : SECRET_ENTER;
+  const entryTokenFound = jwt.verify(request.token, entrySecret).id;
 
   if (!entryTokenFound) {
     return response.status(401).json({error: 'valid token required'});
   }
-  const sceneData = await Scene.find({});
+  const sceneData = await Scene.find({isDemo: isDemo});
 
   if (sceneData) {
     response.json(sceneData);
@@ -44,18 +61,26 @@ sceneRouter.get('/', async (request, response, next) => {
 
 // deleting entire scene (requires ADMIN token)
 sceneRouter.delete('/:id', async (request, response, next) => {
-  const adminTokenFound = jwt.verify(request.token, SECRET_ADMIN).id;
+  const adminTokenFound = request.isDemo ?
+    jwt.verify(request.token, SECRET_ADMIN_DEMO).id :
+    jwt.verify(request.token, SECRET_ADMIN).id;
 
   if (!adminTokenFound) {
     return response.status(401).json({error: 'valid admin required'});
   }
 
   const thisID = request.params.id;
+  const scene = await Scene.findById(thisID);
 
   // scene isn't there
-  const scene = await Scene.findById(thisID);
   if (!scene) {
     return response.status(404).json({error: 'scene not found'});
+  }
+
+  // demo admin can't delete if this is a default scene
+  if (request.isDemo && !scene.isDemo) {
+    return response.status(403)
+        .json({error: 'demo users cannot change default scenes'});
   }
 
   await Scene.findByIdAndDelete(thisID);
